chore(frontend): address `clippy::module_inception` lint and simplify module structure
