--- crates/frontend/src/components/button/mod.rs
@@ -1,1 +1,0 @@
-pub mod button;
