--- crates/frontend/src/components/drawer/mod.rs
@@ -1,1 +1,0 @@
-pub mod drawer;
