--- crates/frontend/src/components/modal/mod.rs
@@ -1,1 +1,0 @@
-pub mod modal;
