+++ crates/frontend/src/components/nav_item.rs
