--- crates/frontend/src/components/nav_item/mod.rs
@@ -1,1 +1,0 @@
-pub mod nav_item;
