+++ crates/frontend/src/components/pagination.rs
