--- crates/frontend/src/components/pagination/mod.rs
@@ -1,1 +1,0 @@
-pub mod pagination;
