--- crates/frontend/src/components/side_nav/mod.rs
@@ -1,1 +1,0 @@
-pub mod side_nav;
