+++ crates/frontend/src/components/stat.rs
