--- crates/frontend/src/components/stat/mod.rs
@@ -1,1 +1,0 @@
-pub mod stat;
