--- crates/frontend/src/components/variant_form/mod.rs
@@ -1,1 +1,0 @@
-pub mod variant_form;
