--- crates/frontend/src/hoc/layout/mod.rs
@@ -1,1 +1,0 @@
-pub mod layout;
