+++ crates/frontend/src/pages/ContextOverride/context_override.rs
@@ -1,14 +1,14 @@
 use crate::api::fetch_config;
 use crate::api::{delete_context, fetch_default_config, fetch_dimensions};
-use crate::components::button::button::Button;
-use crate::components::condition_pills::condition_pills::ConditionPills;
+use crate::components::button::Button;
 use crate::components::condition_pills::types::Condition;
-use crate::components::context_form::context_form::ContextForm;
+use crate::components::condition_pills::ConditionPills;
 use crate::components::context_form::utils::{create_context, update_context};
-use crate::components::drawer::drawer::{close_drawer, open_drawer, Drawer, DrawerBtn};
-use crate::components::override_form::override_form::OverrideForm;
+use crate::components::context_form::ContextForm;
+use crate::components::drawer::{close_drawer, open_drawer, Drawer, DrawerBtn};
+use crate::components::override_form::OverrideForm;
 use crate::components::skeleton::{Skeleton, SkeletonVariant};
-use crate::components::table::{table::Table, types::Column};
+use crate::components::table::{types::Column, Table};
 use crate::types::{Config, DefaultConfig, Dimension};
 use crate::utils::extract_conditions;
 use futures::join;
