--- crates/frontend/src/pages/Dimensions/mod.rs
@@ -1,1 +1,0 @@
-pub mod Dimensions;
