--- crates/frontend/src/pages/Home/mod.rs
@@ -1,2 +1,0 @@
-#![allow(non_snake_case)]
-pub mod Home;
