--- crates/frontend/src/pages/NotFound/mod.rs
@@ -1,1 +1,0 @@
-pub mod NotFound;
