--- crates/frontend/src/pages/default_config/mod.rs
@@ -1,1 +1,0 @@
-pub mod default_config;
